Alteração de tipo de dados

+++ carregar_energia_api/core/models.py
@@ -71,7 +71,7 @@ def report_uploaded(sender, instance, created, **kwards):
 
 class CarregarRecarga(models.Model):
     id_do_usuario = models.ForeignKey(Usuario, on_delete = models.CASCADE)
-    codigo_da_recarga = models.IntegerField()
+    codigo_da_recarga = models.BigIntegerField()
     data_criacao = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
